pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodev"
version = "0.1.0"
description = "An open-source, model-agnostic autonomous engineering runtime"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "typer[all]>=0.9.0",
    "PyGithub>=2.1.1",
    "GitPython>=3.1.40",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
    "PyYAML>=6.0.1",
    "httpx>=0.27.0",
    "openai>=1.12.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
autodev = "autodev.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["autodev*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
